refactor: extract Subscriptions and Orders tables into components
