+++ clients/apps/web/src/components/Products/ProductPage/ProductOverview.tsx
@@ -1,16 +1,8 @@
 import { MiniMetricChartBox } from '@/components/Metrics/MiniMetricChartBox'
-import { OrderStatus } from '@/components/Orders/OrderStatus'
-import { SubscriptionStatus as SubscriptionStatusComponent } from '@/components/Subscriptions/SubscriptionStatus'
-import { useOrders } from '@/hooks/queries/orders'
-import { useSubscriptions } from '@/hooks/queries/subscriptions'
 import { schemas } from '@polar-sh/client'
-import { formatCurrency } from '@polar-sh/currency'
-import { Avatar } from '@polar-sh/orbit'
-import { Button } from '@polar-sh/orbit'
-import { DataTable, DataTableColumnHeader } from '@polar-sh/orbit'
-import FormattedDateTime from '@polar-sh/ui/components/atoms/FormattedDateTime'
-import Link from 'next/link'
 import { ProductDiscounts } from './ProductDiscounts'
+import { ProductOrders } from './ProductOrders'
+import { ProductSubscriptions } from './ProductSubscriptions'
 
 export interface ProductOverviewProps {
   organization: schemas['Organization']
@@ -25,26 +17,6 @@ export const ProductOverview = ({
   metrics,
   todayMetrics,
 }: ProductOverviewProps) => {
-  const { data: productOrders, isLoading: productOrdersIsLoading } = useOrders(
-    organization.id,
-    {
-      product_id: product.id,
-      limit: 10,
-    },
-  )
-
-  const { data: subscriptions, isLoading: subscriptionsIsLoading } =
-    useSubscriptions(
-      product.is_recurring ? organization.id : undefined,
-      product.is_recurring
-        ? {
-            product_id: product.id,
-            active: true,
-            limit: 10,
-          }
-        : undefined,
-    )
-
   return (
     <div className="flex flex-col gap-y-16">
       <div className="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
@@ -80,220 +52,9 @@ export const ProductOverview = ({
         />
       </div>
       {product.is_recurring && (
-        <div className="flex flex-col gap-y-6">
-          <div className="flex flex-row items-center justify-between gap-x-6">
-            <div className="flex flex-col gap-y-1">
-              <h2 className="text-lg">Subscriptions</h2>
-              <p className="dark:text-polar-500 text-sm text-gray-500">
-                Showing 10 most recent subscriptions for {product.name}
-              </p>
-            </div>
-            <Link
-              href={`/dashboard/${organization.slug}/sales/subscriptions?product_id=${product.id}`}
-            >
-              <Button size="sm">View All</Button>
-            </Link>
-          </div>
-          <DataTable
-            data={subscriptions?.items ?? []}
-            columns={[
-              {
-                id: 'customer',
-                accessorKey: 'customer',
-                enableSorting: true,
-                header: ({ column }) => (
-                  <DataTableColumnHeader column={column} title="Customer" />
-                ),
-                cell: ({ row: { original: subscription } }) => {
-                  const customer = subscription.customer
-                  return (
-                    <div className="flex flex-row items-center gap-2">
-                      <Avatar
-                        avatar_url={customer.avatar_url}
-                        name={customer.name ?? customer.email ?? '—'}
-                      />
-                      <div className="flex flex-col overflow-hidden">
-                        <span className="truncate">{customer.name ?? '—'}</span>
-                        <span className="dark:text-polar-500 truncate text-xs text-gray-500">
-                          {customer.email ?? '—'}
-                        </span>
-                      </div>
-                    </div>
-                  )
-                },
-              },
-              {
-                accessorKey: 'status',
-                enableSorting: true,
-                header: ({ column }) => (
-                  <DataTableColumnHeader column={column} title="Status" />
-                ),
-                cell: ({ row: { original: subscription } }) => {
-                  return (
-                    <SubscriptionStatusComponent subscription={subscription} />
-                  )
-                },
-              },
-              {
-                accessorKey: 'started_at',
-                enableSorting: true,
-                header: ({ column }) => (
-                  <DataTableColumnHeader
-                    column={column}
-                    title="Subscription Date"
-                  />
-                ),
-                cell: (props) => (
-                  <FormattedDateTime datetime={props.getValue() as string} />
-                ),
-              },
-              {
-                accessorKey: 'current_period_end',
-                enableSorting: true,
-                header: ({ column }) => (
-                  <DataTableColumnHeader column={column} title="Renewal Date" />
-                ),
-                cell: ({
-                  getValue,
-                  row: {
-                    original: { status, cancel_at_period_end },
-                  },
-                }) => {
-                  const datetime = getValue() as string | null
-                  const willRenew =
-                    (status === 'active' || status === 'trialing') &&
-                    !cancel_at_period_end
-                  return datetime && willRenew ? (
-                    <FormattedDateTime datetime={datetime} />
-                  ) : (
-                    '—'
-                  )
-                },
-              },
-              {
-                accessorKey: 'actions',
-                enableSorting: false,
-                header: () => null,
-                cell: (props) => (
-                  <span className="flex flex-row justify-end gap-x-2">
-                    <Link
-                      href={`/dashboard/${organization.slug}/customers/${props.row.original.customer.id}`}
-                    >
-                      <Button variant="secondary" size="sm">
-                        View Customer
-                      </Button>
-                    </Link>
-                    <Link
-                      href={`/dashboard/${organization.slug}/sales/subscriptions/${props.row.original.id}`}
-                    >
-                      <Button variant="secondary" size="sm">
-                        View Subscription
-                      </Button>
-                    </Link>
-                  </span>
-                ),
-              },
-            ]}
-            isLoading={subscriptionsIsLoading}
-          />
-        </div>
+        <ProductSubscriptions organization={organization} product={product} />
       )}
-      <div className="flex flex-col gap-y-6">
-        <div className="flex flex-row items-center justify-between gap-x-6">
-          <div className="flex flex-col gap-y-1">
-            <h2 className="text-lg">Orders</h2>
-            <p className="dark:text-polar-500 text-sm text-gray-500">
-              Showing last 10 orders for {product.name}
-            </p>
-          </div>
-          <Link
-            href={`/dashboard/${organization.slug}/sales?product_id=${product.id}`}
-          >
-            <Button size="sm">View All</Button>
-          </Link>
-        </div>
-        <DataTable
-          data={productOrders?.items ?? []}
-          columns={[
-            {
-              accessorKey: 'customer',
-              enableSorting: true,
-              header: ({ column }) => (
-                <DataTableColumnHeader column={column} title="Customer" />
-              ),
-              cell: (props) => {
-                const customer = props.getValue() as schemas['OrderCustomer']
-                return (
-                  <div className="flex flex-row items-center gap-2">
-                    <Avatar
-                      className="h-8 w-8"
-                      avatar_url={customer.avatar_url}
-                      name={customer.name ?? customer.email ?? '—'}
-                    />
-                    <div className="flex flex-col overflow-hidden">
-                      <span className="truncate">{customer.name ?? '—'}</span>
-                      <span className="dark:text-polar-500 truncate text-xs text-gray-500">
-                        {customer.email ?? '—'}
-                      </span>
-                    </div>
-                  </div>
-                )
-              },
-            },
-            {
-              accessorKey: 'status',
-              enableSorting: true,
-              header: ({ column }) => (
-                <DataTableColumnHeader column={column} title="Status" />
-              ),
-              cell: ({ row: { original: order } }) => (
-                <span className="flex shrink">
-                  <OrderStatus status={order.status} />
-                </span>
-              ),
-            },
-            {
-              accessorKey: 'amount',
-              enableSorting: true,
-              header: ({ column }) => (
-                <DataTableColumnHeader column={column} title="Amount" />
-              ),
-              cell: ({ row: { original: order } }) => (
-                <span>
-                  {formatCurrency('compact')(order.net_amount, order.currency)}
-                </span>
-              ),
-            },
-            {
-              accessorKey: 'created_at',
-              enableSorting: true,
-              header: ({ column }) => (
-                <DataTableColumnHeader column={column} title="Date" />
-              ),
-              cell: (props) => (
-                <FormattedDateTime datetime={props.getValue() as string} />
-              ),
-            },
-            {
-              accessorKey: 'actions',
-              enableSorting: true,
-              header: () => null,
-              cell: (props) => (
-                <span className="flex flex-row justify-end">
-                  <Link
-                    href={`/dashboard/${organization.slug}/sales/${props.row.original.id}`}
-                  >
-                    <Button variant="secondary" size="sm">
-                      View
-                    </Button>
-                  </Link>
-                </span>
-              ),
-            },
-          ]}
-          isLoading={productOrdersIsLoading}
-        />
-      </div>
+      <ProductOrders organization={organization} product={product} />
 
       {!product.is_archived && (
         <ProductDiscounts organization={organization} product={product} />
